Rename orders subpackage
