+++ examples/strategies/ema_cross_simple.py
@@ -25,8 +25,8 @@ from nautilus_trader.model.enums import OrderSide
 from nautilus_trader.model.identifiers import InstrumentId
 from nautilus_trader.model.instrument import Instrument
 from nautilus_trader.model.objects import Quantity
-from nautilus_trader.model.order.market import MarketOrder
 from nautilus_trader.model.orderbook.book import OrderBook
+from nautilus_trader.model.orders.market import MarketOrder
 from nautilus_trader.model.tick import QuoteTick
 from nautilus_trader.model.tick import TradeTick
 from nautilus_trader.trading.strategy import TradingStrategy
