+++ examples/strategies/ema_cross_stop_entry_trail.py
@@ -28,8 +28,8 @@ from nautilus_trader.model.identifiers import InstrumentId
 from nautilus_trader.model.instrument import Instrument
 from nautilus_trader.model.objects import Price
 from nautilus_trader.model.objects import Quantity
-from nautilus_trader.model.order.stop_market import StopMarketOrder
 from nautilus_trader.model.orderbook.book import OrderBook
+from nautilus_trader.model.orders.stop_market import StopMarketOrder
 from nautilus_trader.model.tick import QuoteTick
 from nautilus_trader.model.tick import TradeTick
 from nautilus_trader.trading.strategy import TradingStrategy
