+++ nautilus_trader/adapters/betfair/parsing.py
@@ -65,11 +65,11 @@ from nautilus_trader.model.instrument import BettingInstrument
 from nautilus_trader.model.objects import Money
 from nautilus_trader.model.objects import Price
 from nautilus_trader.model.objects import Quantity
-from nautilus_trader.model.order.limit import LimitOrder
 from nautilus_trader.model.orderbook.book import OrderBookDelta
 from nautilus_trader.model.orderbook.book import OrderBookDeltas
 from nautilus_trader.model.orderbook.book import OrderBookSnapshot
 from nautilus_trader.model.orderbook.order import Order
+from nautilus_trader.model.orders.limit import LimitOrder
 from nautilus_trader.model.tick import TradeTick
