+++ nautilus_trader/adapters/ccxt/execution.pxd
@@ -15,7 +15,7 @@
 
 from nautilus_trader.live.execution_client cimport LiveExecutionClient
 from nautilus_trader.model.identifiers cimport VenueOrderId
-from nautilus_trader.model.order.base cimport Order
+from nautilus_trader.model.orders.base cimport Order
 
 
 cdef class CCXTExecutionClient(LiveExecutionClient):
