+++ nautilus_trader/model/orders/limit.pxd
@@ -14,12 +14,16 @@
 # -------------------------------------------------------------------------------------------------
 
 from nautilus_trader.model.events cimport OrderInitialized
-from nautilus_trader.model.order.base cimport PassiveOrder
+from nautilus_trader.model.orders.base cimport PassiveOrder
 
 
-cdef class StopMarketOrder(PassiveOrder):
+cdef class LimitOrder(PassiveOrder):
+    cdef readonly bint is_post_only
+    """If the order will only make liquidity.\n\n:returns: `bool`"""
     cdef readonly bint is_reduce_only
     """If the order will only reduce an open position.\n\n:returns: `bool`"""
+    cdef readonly bint is_hidden
+    """If the order is hidden from the public book.\n\n:returns: `bool`"""
 
     @staticmethod
-    cdef StopMarketOrder create(OrderInitialized init)
+    cdef LimitOrder create(OrderInitialized init)
