+++ nautilus_trader/model/orders/market.pxd
@@ -13,17 +13,11 @@
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
 
+from nautilus_trader.core.constants cimport *  # str constants only
 from nautilus_trader.model.events cimport OrderInitialized
-from nautilus_trader.model.order.base cimport PassiveOrder
+from nautilus_trader.model.orders.base cimport Order
 
 
-cdef class LimitOrder(PassiveOrder):
-    cdef readonly bint is_post_only
-    """If the order will only make liquidity.\n\n:returns: `bool`"""
-    cdef readonly bint is_reduce_only
-    """If the order will only reduce an open position.\n\n:returns: `bool`"""
-    cdef readonly bint is_hidden
-    """If the order is hidden from the public book.\n\n:returns: `bool`"""
-
+cdef class MarketOrder(Order):
     @staticmethod
-    cdef LimitOrder create(OrderInitialized init)
+    cdef MarketOrder create(OrderInitialized init)
