+++ nautilus_trader/model/orders/stop_limit.pxd
@@ -13,11 +13,22 @@
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
 
-from nautilus_trader.core.constants cimport *  # str constants only
 from nautilus_trader.model.events cimport OrderInitialized
-from nautilus_trader.model.order.base cimport Order
+from nautilus_trader.model.objects cimport Price
+from nautilus_trader.model.orders.base cimport PassiveOrder
 
 
-cdef class MarketOrder(Order):
+cdef class StopLimitOrder(PassiveOrder):
+    cdef readonly Price trigger
+    """The trigger stop price for the order.\n\n:returns: `Price`"""
+    cdef readonly bint is_triggered
+    """If the order has been triggered.\n\n:returns: `bool`"""
+    cdef readonly bint is_post_only
+    """If the order will only make liquidity.\n\n:returns: `bool`"""
+    cdef readonly bint is_reduce_only
+    """If the order will only reduce an open position.\n\n:returns: `bool`"""
+    cdef readonly bint is_hidden
+    """If the order is hidden from the public book.\n\n:returns: `bool`"""
+
     @staticmethod
-    cdef MarketOrder create(OrderInitialized init)
+    cdef StopLimitOrder create(OrderInitialized init)
