+++ nautilus_trader/model/orders/stop_limit.pyx
@@ -23,21 +23,31 @@ from nautilus_trader.model.c_enums.order_side cimport OrderSide
 from nautilus_trader.model.c_enums.order_type cimport OrderType
 from nautilus_trader.model.c_enums.time_in_force cimport TimeInForce
 from nautilus_trader.model.events cimport OrderInitialized
+from nautilus_trader.model.events cimport OrderTriggered
+from nautilus_trader.model.events cimport OrderUpdated
 from nautilus_trader.model.identifiers cimport ClientOrderId
 from nautilus_trader.model.identifiers cimport InstrumentId
 from nautilus_trader.model.identifiers cimport StrategyId
 from nautilus_trader.model.objects cimport Price
 from nautilus_trader.model.objects cimport Quantity
-from nautilus_trader.model.order.base cimport PassiveOrder
+from nautilus_trader.model.orders.base cimport PassiveOrder
 
 
-cdef class LimitOrder(PassiveOrder):
+cdef class StopLimitOrder(PassiveOrder):
     """
-    Limit orders are used to specify a maximum or minimum price the trader is
-    willing to buy or sell at. Traders use this order type to minimise their
-    trading cost, however they are sacrificing guaranteed execution as there is
-    a chance the order may not be executed if it is placed deep out of the
-    market.
+    Represents a stop-limit order.
+
+    A stop-limit order is an instruction to submit a buy or sell limit order
+    when the user-specified stop trigger price is attained or penetrated. The
+    order has two basic components: the stop price and the limit price. When a
+    trade has occurred at or through the stop price, the order becomes
+    executable and enters the market as a limit order, which is an order to buy
+    or sell at a specified price or better.
+
+    A stop-limit eliminates the price risk associated with a stop order where
+    the execution price cannot be guaranteed, but exposes the trader to the
+    risk that the order may never fill even if the stop price is reached. The
+    trader could "miss the market" altogether.
     """
     def __init__(
         self,
@@ -47,6 +57,7 @@ cdef class LimitOrder(PassiveOrder):
         OrderSide order_side,
         Quantity quantity not None,
         Price price not None,
+        Price trigger not None,
         TimeInForce time_in_force,
         datetime expire_time,  # Can be None
         UUID init_id not None,
@@ -56,7 +67,7 @@ cdef class LimitOrder(PassiveOrder):
         bint hidden=False,
     ):
         """
-        Initialize a new instance of the `LimitOrder` class.
+        Initialize a new instance of the `StopLimitOrder` class.
 
         Parameters
         ----------
@@ -72,6 +83,8 @@ cdef class LimitOrder(PassiveOrder):
             The order quantity (> 0).
         price : Price
             The order limit price.
+        trigger : Price
+            The order stop trigger price.
         time_in_force : TimeInForce
             The order time-in-force.
         expire_time : datetime, optional
@@ -81,22 +94,19 @@ cdef class LimitOrder(PassiveOrder):
         timestamp_ns : int64
             The Unix timestamp (nanos) of the order initialization.
         post_only : bool, optional
-            If the order will only make a market.
+            If the order will only make a market (once triggered).
         reduce_only : bool, optional
-            If the order will only reduce an open position.
+            If the order will only reduce an open position (once triggered).
         hidden : bool, optional
-            If the order will be hidden from the public book.
+            If the order will be hidden from the public book (once triggered).
+
 
         Raises
         ------
         ValueError
             If quantity is not positive (> 0).
         ValueError
-            If time_in_force is GTD and expire_time is None.
-        ValueError
-            If post_only and hidden.
-        ValueError
-            If hidden and post_only.
+            If time_in_force is GTD and the expire_time is None.
 
         """
         if post_only:
@@ -108,7 +118,7 @@ cdef class LimitOrder(PassiveOrder):
             strategy_id=strategy_id,
             instrument_id=instrument_id,
             order_side=order_side,
-            order_type=OrderType.LIMIT,
+            order_type=OrderType.STOP_LIMIT,
             quantity=quantity,
             price=price,
             time_in_force=time_in_force,
@@ -116,20 +126,32 @@ cdef class LimitOrder(PassiveOrder):
             init_id=init_id,
             timestamp_ns=timestamp_ns,
             options={
+                TRIGGER: str(trigger),
                 POST_ONLY: post_only,
                 REDUCE_ONLY: reduce_only,
                 HIDDEN: hidden,
             },
         )
 
+        self.trigger = trigger
+        self.is_triggered = False
         self.is_post_only = post_only
         self.is_reduce_only = reduce_only
         self.is_hidden = hidden
 
+    def __repr__(self) -> str:
+        cdef str id_string = f", id={self.venue_order_id.value})" if self.venue_order_id.not_null() else ")"
+        return (f"{type(self).__name__}("
+                f"{self.status_string_c()}, "
+                f"trigger={self.trigger}, "
+                f"state={self._fsm.state_string_c()}, "
+                f"client_order_id={self.client_order_id.value}"
+                f"{id_string}")
+
     @staticmethod
-    cdef LimitOrder create(OrderInitialized init):
+    cdef StopLimitOrder create(OrderInitialized init):
         """
-        Return a limit order from the given initialized event.
+        Return a stop-limit order from the given initialized event.
 
         Parameters
         ----------
@@ -138,24 +160,25 @@ cdef class LimitOrder(PassiveOrder):
 
         Returns
         -------
-        LimitOrder
+        StopLimitOrder
 
         Raises
         ------
         ValueError
-            If init.order_type is not equal to LIMIT.
+            If init.order_type is not equal to STOP_LIMIT.
 
         """
         Condition.not_none(init, "init")
-        Condition.equal(init.order_type, OrderType.LIMIT, "init.order_type", "OrderType")
+        Condition.equal(init.order_type, OrderType.STOP_LIMIT, "init.order_type", "OrderType")
 
-        return LimitOrder(
+        return StopLimitOrder(
             client_order_id=init.client_order_id,
             strategy_id=init.strategy_id,
             instrument_id=init.instrument_id,
             order_side=init.order_side,
             quantity=init.quantity,
             price=Price(init.options[PRICE]),
+            trigger=Price(init.options[TRIGGER]),
             time_in_force=init.time_in_force,
             expire_time=init.options.get(EXPIRE_TIME),
             init_id=init.id,
@@ -164,3 +187,14 @@ cdef class LimitOrder(PassiveOrder):
             reduce_only=init.options[REDUCE_ONLY],
             hidden=init.options[HIDDEN],
         )
+
+    cdef void _updated(self, OrderUpdated event) except *:
+        self.venue_order_id = event.venue_order_id
+        self.quantity = event.quantity
+        if self.is_triggered:
+            self.price = event.price
+        else:
+            self.trigger = event.price
+
+    cdef void _triggered(self, OrderTriggered event) except *:
+        self.is_triggered = True
