+++ nautilus_trader/model/orders/stop_market.pyx
@@ -13,45 +13,36 @@
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
 
+from cpython.datetime cimport datetime
 from libc.stdint cimport int64_t
 
 from nautilus_trader.core.constants cimport *  # str constants only
 from nautilus_trader.core.correctness cimport Condition
 from nautilus_trader.core.uuid cimport UUID
 from nautilus_trader.model.c_enums.order_side cimport OrderSide
-from nautilus_trader.model.c_enums.order_side cimport OrderSideParser
 from nautilus_trader.model.c_enums.order_type cimport OrderType
-from nautilus_trader.model.c_enums.order_type cimport OrderTypeParser
 from nautilus_trader.model.c_enums.time_in_force cimport TimeInForce
-from nautilus_trader.model.c_enums.time_in_force cimport TimeInForceParser
-from nautilus_trader.model.events cimport OrderFilled
 from nautilus_trader.model.events cimport OrderInitialized
-from nautilus_trader.model.events cimport OrderUpdated
 from nautilus_trader.model.identifiers cimport ClientOrderId
 from nautilus_trader.model.identifiers cimport InstrumentId
 from nautilus_trader.model.identifiers cimport StrategyId
+from nautilus_trader.model.objects cimport Price
 from nautilus_trader.model.objects cimport Quantity
-from nautilus_trader.model.order.base cimport Order
+from nautilus_trader.model.orders.base cimport PassiveOrder
 
 
-cdef set _MARKET_ORDER_VALID_TIF = {
-    TimeInForce.GTC,
-    TimeInForce.IOC,
-    TimeInForce.FOK,
-    TimeInForce.FAK,
-}
-
-
-cdef class MarketOrder(Order):
+cdef class StopMarketOrder(PassiveOrder):
     """
-    Represents a market order.
+    Represents a stop-market order.
 
-    A market order is an order to buy or sell an instrument immediately. This
-    type of order guarantees that the order will be executed, but does not
-    guarantee the execution price. A market order generally will execute at or
-    near the current bid (for a sell order) or ask (for a buy order) price. The
-    last-traded price is not necessarily the price at which a market order will
-    be executed.
+    A stop-market order is an instruction to submit a buy or sell market order
+    if and when the user-specified stop trigger price is attained or penetrated.
+    A stop-market order is not guaranteed a specific execution price and may
+    execute significantly away from its stop price. A Sell Stop order is always
+    placed below the current market price and is typically used to limit a loss
+    or protect a profit on a long stock position. A Buy Stop order is always
+    placed above the current market price. It is typically used to limit a loss
+    or help protect a profit on a short sale.
     """
     def __init__(
         self,
@@ -60,12 +51,15 @@ cdef class MarketOrder(Order):
         InstrumentId instrument_id not None,
         OrderSide order_side,
         Quantity quantity not None,
+        Price price not None,
         TimeInForce time_in_force,
+        datetime expire_time,  # Can be None
         UUID init_id not None,
         int64_t timestamp_ns,
+        bint reduce_only=False,
     ):
         """
-        Initialize a new instance of the `MarketOrder` class.
+        Initialize a new instance of the `StopMarketOrder` class.
 
         Parameters
         ----------
@@ -79,41 +73,48 @@ cdef class MarketOrder(Order):
             The order side (BUY or SELL).
         quantity : Quantity
             The order quantity (> 0).
+        price : Price
+            The order stop price.
+        time_in_force : TimeInForce
+            The order time-in-force.
+        expire_time : datetime, optional
+            The order expiry time.
         init_id : UUID
             The order initialization event identifier.
         timestamp_ns : int64
             The Unix timestamp (nanos) of the order initialization.
+        reduce_only : bool, optional
+            If the order will only reduce an open position.
 
         Raises
         ------
         ValueError
             If quantity is not positive (> 0).
         ValueError
-            If time_in_force is other than GTC, IOC or FOK.
+            If time_in_force is GTD and the expire_time is None.
 
         """
-        Condition.positive(quantity, "quantity")
-        Condition.true(time_in_force in _MARKET_ORDER_VALID_TIF, "time_in_force was != GTC, IOC or FOK")
-
-        cdef OrderInitialized init = OrderInitialized(
+        super().__init__(
             client_order_id=client_order_id,
             strategy_id=strategy_id,
             instrument_id=instrument_id,
             order_side=order_side,
-            order_type=OrderType.MARKET,
+            order_type=OrderType.STOP_MARKET,
             quantity=quantity,
+            price=price,
             time_in_force=time_in_force,
-            event_id=init_id,
+            expire_time=expire_time,
+            init_id=init_id,
             timestamp_ns=timestamp_ns,
-            options={},
+            options={REDUCE_ONLY: reduce_only},
         )
 
-        super().__init__(init=init)
+        self.is_reduce_only = reduce_only
 
     @staticmethod
-    cdef MarketOrder create(OrderInitialized init):
+    cdef StopMarketOrder create(OrderInitialized init):
         """
-        Return an order from the given initialized event.
+        Return a stop-market order from the given initialized event.
 
         Parameters
         ----------
@@ -122,42 +123,27 @@ cdef class MarketOrder(Order):
 
         Returns
         -------
-        MarketOrder
+        StopMarketOrder
 
         Raises
         ------
         ValueError
-            If init.order_type is not equal to MARKET.
+            If init.order_type is not equal to STOP_MARKET.
 
         """
         Condition.not_none(init, "init")
-        Condition.equal(init.order_type, OrderType.MARKET, "init.order_type", "OrderType")
+        Condition.equal(init.order_type, OrderType.STOP_MARKET, "init.order_type", "OrderType")
 
-        return MarketOrder(
+        return StopMarketOrder(
             client_order_id=init.client_order_id,
             strategy_id=init.strategy_id,
             instrument_id=init.instrument_id,
             order_side=init.order_side,
             quantity=init.quantity,
+            price=Price(init.options[PRICE]),
             time_in_force=init.time_in_force,
+            expire_time=init.options.get(EXPIRE_TIME),
             init_id=init.id,
             timestamp_ns=init.timestamp_ns,
+            reduce_only=init.options[REDUCE_ONLY],
         )
-
-    cdef str status_string_c(self):
-        return (f"{OrderSideParser.to_str(self.side)} {self.quantity.to_str()} {self.instrument_id} "
-                f"{OrderTypeParser.to_str(self.type)} "
-                f"{TimeInForceParser.to_str(self.time_in_force)}")
-
-    cdef void _updated(self, OrderUpdated event) except *:
-        raise NotImplemented("Cannot update a market order")
-
-    cdef void _filled(self, OrderFilled fill) except *:
-        self.venue_order_id = fill.venue_order_id
-        self.position_id = fill.position_id
-        self.strategy_id = fill.strategy_id
-        self._execution_ids.append(fill.execution_id)
-        self.execution_id = fill.execution_id
-        self.filled_qty = Quantity(self.filled_qty + fill.last_qty)
-        self.execution_ns = fill.execution_ns
-        self.avg_px = self._calculate_avg_px(fill.last_qty, fill.last_px)
