+++ nautilus_trader/serialization/base.pyx
@@ -16,7 +16,7 @@
 from nautilus_trader.core.message cimport Command
 from nautilus_trader.core.message cimport Event
 from nautilus_trader.model.instrument cimport Instrument
-from nautilus_trader.model.order.base cimport Order
+from nautilus_trader.model.orders.base cimport Order
 
 
 cdef class Serializer:
