+++ tests/unit_tests/serialization/test_serialization_serializers.py
@@ -51,9 +51,9 @@ from nautilus_trader.model.identifiers import VenueOrderId
 from nautilus_trader.model.objects import Money
 from nautilus_trader.model.objects import Price
 from nautilus_trader.model.objects import Quantity
-from nautilus_trader.model.order.limit import LimitOrder
-from nautilus_trader.model.order.stop_limit import StopLimitOrder
-from nautilus_trader.model.order.stop_market import StopMarketOrder
+from nautilus_trader.model.orders.limit import LimitOrder
+from nautilus_trader.model.orders.stop_limit import StopLimitOrder
+from nautilus_trader.model.orders.stop_market import StopMarketOrder
 from nautilus_trader.serialization.base import Serializer
 from nautilus_trader.serialization.serializers import MsgPackCommandSerializer
 from nautilus_trader.serialization.serializers import MsgPackEventSerializer
